Refactor: Http.getItems take escaping closure
* Move takeAsseData(assetName:) to SessionMock
* Fix typo
