+++ OpenMarket/ParsingTest/ParsingTest.swift
@@ -27,13 +27,13 @@ class ParsingTest: XCTestCase {
         let givenURL = baseURL + "blah blah"
         let expectedResult = HttpConfig.invailedPath
         
-        let result = sessionMock.httpGetItems(url: givenURL)
-        
-        switch result {
-        case .success(let itemList):
-            XCTAssertEqual("check", itemList.items.description)
-        case .failure(let error):
-            XCTAssertEqual(expectedResult, error.message)
+        sessionMock.getItems(from: givenURL) { result in
+            switch result {
+            case .success(let itemList):
+                XCTAssertEqual("check", itemList.items.description)
+            case .failure(let error):
+                XCTAssertEqual(expectedResult, error.message)
+            }
         }
     }
     
@@ -41,39 +41,39 @@ class ParsingTest: XCTestCase {
         let givenURL = baseURL + "/itesm/-1"
         let expectedResult = HttpConfig.invailedPath
         
-        let result = sessionMock.httpGetItems(url: givenURL)
-        
-        switch result {
-        case .success(let itemList):
-            XCTAssertEqual("check", itemList.items.description)
-        case .failure(let error):
-            XCTAssertEqual(expectedResult, error.message)
+        sessionMock.getItems(from: givenURL) { result in
+            switch result {
+            case .success(let itemList):
+                XCTAssertEqual("check", itemList.items.description)
+            case .failure(let error):
+                XCTAssertEqual(expectedResult, error.message)
+            }
         }
     }
     
     func test_정상적인url로_items에_1번page로_접근하면_아이템목록이있다() {
         let givenURL = baseURL + "/items/1"
         
-        let result = sessionMock.httpGetItems(url: givenURL)
-        
-        switch result {
-        case .success(let itemList):
-            XCTAssertTrue(itemList.items.count > 0)
-        case .failure(let error):
-            XCTAssertEqual("check", error.message)
+        sessionMock.getItems(from: givenURL) { result in
+            switch result {
+            case .success(let itemList):
+                XCTAssertTrue(itemList.items.count > 0)
+            case .failure(let error):
+                XCTAssertEqual("check", error.message)
+            }
         }
     }
     
     func test_정상적인url로_items에_100번page로_접근하면_아이템목록이없다() {
         let givenURL = baseURL + "/items/100"
         
-        let result = sessionMock.httpGetItems(url: givenURL)
-        
-        switch result {
-        case .success(let itemList):
-            XCTAssertTrue(itemList.items.count == 0)
-        case .failure(let error):
-            XCTAssertEqual("check", error.message)
+        sessionMock.getItems(from: givenURL) { result in
+            switch result {
+            case .success(let itemList):
+                XCTAssertTrue(itemList.items.count == 0)
+            case .failure(let error):
+                XCTAssertEqual("check", error.message)
+            }
         }
     }
